pyproject.toml
[project]
name = "verdes"
version = "1"
description = "Stardew Valley benzeri yapay zeka destekli çiftlik oyunu"
readme = "README.rst"
authors = [
  {name = "Verde", email = "berkan.baser01@gmail.com"}
]
maintainers = [
  {name = "Verde", email = "berkan.baser01@gmail.com"}
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Games/Entertainment :: Simulation",
]
license = {text = "MIT license"}
dependencies = [
  "typer",
  "pygame>=2.5.0",
  "pgzero>=1.2.1",
  "numpy>=1.25.0",
  "torch>=2.0.0",  # CPU sürümü
  "transformers>=4.30.0",
  "pyyaml>=6.0",
  "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "coverage",
    "mypy",
    "pytest",
    "ruff"
]